feat(controllers): add controllers

+++ src/main/resources/db/schema.sql
@@ -22,10 +22,10 @@ CREATE TABLE club (
 CREATE TABLE player (
     id UUID PRIMARY KEY,
     name VARCHAR(128) NOT NULL,
-    number INTEGER,
+    number BIGINT,
     position VARCHAR(16) CHECK (position IN ('STRIKER', 'MIDFIELDER', 'DEFENSE', 'GOAL_KEEPER')),
     nationality VARCHAR(50),
-    age INTEGER CHECK (age BETWEEN 0 and 80),
+    age BIGINT CHECK (age BETWEEN 0 and 80),
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
 
@@ -34,7 +34,7 @@ CREATE TABLE player (
 
 CREATE TABLE season (
     id UUID PRIMARY KEY,
-    year INTEGER NOT NULL,
+    year BIGINT NOT NULL,
     alias VARCHAR(24) NOT NULL,
     status VARCHAR(16) NOT NULL CHECK (status IN ('NOT_STARTED', 'STARTED', 'FINISHED'))
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -46,8 +46,8 @@ CREATE TABLE "match" (
     stadium VARCHAR(100),
     match_datetime TIMESTAMP NOT NULL,
     status VARCHAR(20) NOT NULL CHECK (sta tus IN ('NOT_STARTED', 'STARTED', 'FINISHED')),
-    score_home INTEGER DEFAULT 0,
-    score_away INTEGER DEFAULT 0,
+    score_home BIGINT DEFAULT 0,
+    score_away BIGINT DEFAULT 0,
 
     season_id UUID NOT NULL REFERENCES season(id),
     club_home_id UUID NOT NULL REFERENCES club(id),
@@ -64,7 +64,7 @@ CREATE TABLE match_player (
 
 CREATE TABLE goal (
     id SERIAL PRIMARY KEY,
-    minute INTEGER NOT NULL CHECK (minute BETWEEN 1 AND 90),
+    minute BIGINT NOT NULL CHECK (minute BETWEEN 1 AND 90),
     own_goal BOOLEAN NOT NULL DEFAULT FALSE,
 
     match_id UUID NOT NULL REFERENCES "match"(id),
@@ -74,8 +74,8 @@ CREATE TABLE goal (
 );
 
 CREATE TABLE player_statistics (
-    scored_goals INTEGER DEFAULT 0,
-    playing_time_value NUMERIC,
+    scored_goals BIGINT DEFAULT 0,
+    playing_time_value BIGINT,
     playing_time_unit VARCHAR(12) CHECK (playing_time_unit IN ('SECOND', 'MINUTE', 'HOUR')),
 
     player_id UUID NOT NULL REFERENCES player(id),
@@ -84,11 +84,11 @@ CREATE TABLE player_statistics (
 );
 
 CREATE TABLE club_statistics (
-    ranking_points INTEGER DEFAULT 0,
-    scored_goals INTEGER DEFAULT 0,
-    conceded_goals INTEGER DEFAULT 0,
-    difference_goals INTEGER DEFAULT 0,
-    clean_sheet_number INTEGER DEFAULT 0,
+    ranking_points BIGINT DEFAULT 0,
+    scored_goals BIGINT DEFAULT 0,
+    conceded_goals BIGINT DEFAULT 0,
+    difference_goals BIGINT DEFAULT 0,
+    clean_sheet_number BIGINT DEFAULT 0,
 
     club_id UUID NOT NULL REFERENCES club(id),
     season_id UUID NOT NULL REFERENCES season(id),
